requirements.txt
altair==4.2.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
fonttools==4.39.2
gitdb==4.0.10
GitPython==3.1.31
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.1.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
multimethod==1.9.1
networkx==3.0
numpy==1.23.5
packaging==23.0
pandas==1.5.3
pandas-profiling==3.6.6
patsy==0.5.3
phik==0.12.3
Pillow==9.4.0
plotly==5.14.1
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
rich==13.3.2
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.12.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
statsmodels==0.13.5
streamlit==1.20.0
streamlit-pandas-profiling==0.1.3
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
visions==0.7.5
watchdog==3.0.0
ydata-profiling==4.1.1
zipp==3.15.0
